+++ um_members.py
@@ -10,7 +10,7 @@ from Login.verification import Verification
 from SuperAdmin import super_admin_menu as SuperMenu
 from systemAdmin import system_admin_menu as SystemMenu
 from serviceEngineer import ServiceEngineer_menu
-from DbContext.backup_utils import create_backup, list_backups, restore_backup, delete_backup
+from backup.backup_menu import backup_menu
 from systemAdmin.system_admin import systemAdmin
 
 DB_PATH = "data.db"
@@ -189,308 +189,6 @@ def show_main_menu(role, username):
     else:
         print("Invalid role.")
         return
-
-def generate_restore_code(length=12):
-    """Generate a random alphanumeric restore code."""
-    return ''.join(random.choices(string.ascii_uppercase + string.digits, k=length))
-
-def add_restore_code(backup_name, system_admin, db_path=DB_PATH):
-    # Fetch all entries and decrypt to check for existing unused code
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("SELECT backup_name, system_admin, used FROM backup_recovery_list")
-    rows = cursor.fetchall()
-    for enc_backup_name, enc_system_admin, used in rows:
-        try:
-            dec_backup_name = decrypt(enc_backup_name)
-            dec_system_admin = decrypt(enc_system_admin)
-        except Exception:
-            continue
-        if dec_backup_name == backup_name and dec_system_admin == system_admin and used == 0:
-            conn.close()
-            print(f"System Admin '{system_admin}' already has an active recovery code for backup '{backup_name}'.")
-            return None
-    # Generate a new restore code
-    code = generate_restore_code()
-    enc_backup_name = encrypt(backup_name)
-    enc_system_admin = encrypt(system_admin)
-    enc_code = encrypt(code)
-    cursor.execute("""
-        INSERT INTO backup_recovery_list (backup_name, system_admin, recovery_code, used)
-        VALUES (?, ?, ?, 0)
-    """, (enc_backup_name, enc_system_admin, enc_code))
-    conn.commit()
-    conn.close()
-    return code
-
-def revoke_restore_code(backup_name, system_admin, db_path=DB_PATH):
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("SELECT id, backup_name, system_admin, used FROM backup_recovery_list")
-    rows = cursor.fetchall()
-    for row in rows:
-        row_id, enc_backup_name, enc_system_admin, used = row
-        if used == 0 and decrypt(enc_backup_name) == backup_name and decrypt(enc_system_admin) == system_admin:
-            cursor.execute("""
-                UPDATE backup_recovery_list
-                SET used = 1, used_at = datetime('now')
-                WHERE id = ?
-            """, (row_id,))
-            conn.commit()
-            conn.close()
-            return True
-    conn.close()
-    return False
-
-def validate_restore_code(backup_name, system_admin, code, db_path=DB_PATH):
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("SELECT id, backup_name, system_admin, recovery_code, used FROM backup_recovery_list")
-    rows = cursor.fetchall()
-    for row in rows:
-        row_id, enc_backup_name, enc_system_admin, enc_code, used = row
-        if (used == 0 and decrypt(enc_backup_name) == backup_name and
-            decrypt(enc_system_admin) == system_admin and decrypt(enc_code) == code):
-            cursor.execute("""
-                UPDATE backup_recovery_list
-                SET used = 1, used_at = datetime('now')
-                WHERE id = ?
-            """, (row_id,))
-            conn.commit()
-            conn.close()
-            return True
-    conn.close()
-    return False
-
-def get_system_admins(db_path=DB_PATH):
-    conn = sqlite3.connect(db_path)
-    cursor = conn.cursor()
-    cursor.execute("SELECT Username FROM User WHERE Role = 'systemadmin' AND IsActive = 1")
-    admins = [decrypt(row[0]) for row in cursor.fetchall()]
-    conn.close()
-    return admins
-
-def get_decrypted_backups():
-    backups = list_backups()
-    # backups on disk are not encrypted, but in DB they are
-    return backups
-
-def backup_menu(role, username=None):
-    logger = EncryptedLogger()
-    while True:
-        print("\n=== BACKUP & RESTORE MENU ===")
-        print("1. Create Backup")
-        print("2. List Backups")
-        print("3. Restore Backup")
-        print("4. Delete Backup")
-        if role == "superadmin":
-            print("5. Generate Restore-Code for System Admin")
-            print("6. Revoke Restore-Code")
-            print("7. Exit Backup Menu")
-            valid_choices = ["1", "2", "3", "4", "5", "6", "7"]
-        else:
-            print("5. Exit Backup Menu")
-            valid_choices = ["1", "2", "3", "4", "5"]
-        choice = input("Enter your choice: ").strip()
-        if choice not in valid_choices:
-            print("Invalid choice. Please enter a valid option.")
-            logger.log_entry(username or "system", "Backup Menu", f"Invalid menu choice: {choice}", "No")
-            continue
-        if choice == "1":
-            backup_path = create_backup(username)
-            print(f"Backup created: {backup_path}")
-            logger.log_entry(username or "system", "Backup Menu", f"Created backup: {backup_path}", "No")
-            if role == "systemadmin":
-                print("please contact a Super Admin to add a recovery code for this backup.")
-            if role == "superadmin":
-                while True:
-                    add_code = input("Do you want to add a recovery code for a System Admin? (yes/no): ").strip().lower()
-                    if add_code not in ["yes", "no"]:
-                        print("Invalid input. Please enter 'yes' or 'no'.")
-                        logger.log_entry(username or "system", "Backup Menu", f"Invalid input for add recovery code: {add_code}", "No")
-                        continue
-                    if add_code == "no":
-                        logger.log_entry(username or "system", "Backup Menu", "Chose not to add recovery code after backup.", "No")
-                        break
-                    admins = get_system_admins()
-                    if not admins:
-                        print("No active System Admins found.")
-                        logger.log_entry(username or "system", "Backup Menu", "No active System Admins found for recovery code.", "No")
-                        break
-                    while True:
-                        print("System Admins:")
-                        for idx, admin in enumerate(admins, 1):
-                            print(f"{idx}. {admin}")
-                        sel = input("Select System Admin number: ").strip()
-                        if not sel.isdigit() or int(sel) < 1 or int(sel) > len(admins):
-                            print("Invalid selection. Please enter a valid number.")
-                            logger.log_entry(username or "system", "Backup Menu", f"Invalid System Admin selection: {sel}", "No")
-                            continue
-                        sel_idx = int(sel) - 1
-                        try:
-                            code = add_restore_code(os.path.basename(backup_path), admins[sel_idx])
-                            if code:
-                                print(f"Restore code for {admins[sel_idx]} and backup {os.path.basename(backup_path)}: {code}")
-                                logger.log_entry(username or "system", "Backup Menu", f"Generated restore code for {admins[sel_idx]} and backup {os.path.basename(backup_path)}", "No")
-                            break
-                        except Exception as e:
-                            print(f"Failed to generate restore code: {e}")
-                            logger.log_entry(username or "system", "Backup Menu", f"Failed to generate restore code: {e}", "Yes")
-                    break
-        elif choice == "2":
-            backups = list_backups()
-            logger.log_entry(username or "system", "Backup Menu", "Listed backups", "No")
-            if backups:
-                print("Available backups:")
-                for b in backups:
-                    print(f"- {b}")
-            else:
-                print("No backups found.")
-        elif choice == "3":
-            backups = list_backups()
-            if not backups:
-                print("No backups to restore.")
-                logger.log_entry(username or "system", "Backup Menu", "No backups to restore.", "No")
-                continue
-            while True:
-                print("Available backups:")
-                for idx, b in enumerate(backups, 1):
-                    print(f"{idx}. {b}")
-                sel = input("Select backup number to restore: ").strip()
-                if not sel.isdigit() or int(sel) < 1 or int(sel) > len(backups):
-                    print("Invalid selection. Please enter a valid number.")
-                    logger.log_entry(username or "system", "Backup Menu", f"Invalid restore selection: {sel}", "No")
-                    continue
-                sel_idx = int(sel) - 1
-                if role == "systemadmin":
-                    code = input("Enter your restore code: ").strip()
-                    if not validate_restore_code(backups[sel_idx], username, code):
-                        print("Invalid or already used restore code.")
-                        logger.log_entry(username or "system", "Backup Menu", f"Invalid or used restore code for {backups[sel_idx]}", "Yes")
-                        continue
-                try:
-                    restore_backup(backups[sel_idx], username)
-                    print("Restore complete. Please restart the application.")
-                    logger.log_entry(username or "system", "Backup Menu", f"Restored backup: {backups[sel_idx]}", "No")
-                    exit()
-                except Exception as e:
-                    print(f"Restore failed: {e}")
-                    logger.log_entry(username or "system", "Backup Menu", f"Restore failed: {e}", "Yes")
-                break
-        elif choice == "4":
-            backups = list_backups()
-            if not backups:
-                print("No backups to delete.")
-                logger.log_entry(username or "system", "Backup Menu", "No backups to delete.", "No")
-                continue
-            while True:
-                print("Available backups:")
-                for idx, b in enumerate(backups, 1):
-                    print(f"{idx}. {b}")
-                sel = input("Select backup number to delete: ").strip()
-                if not sel.isdigit() or int(sel) < 1 or int(sel) > len(backups):
-                    print("Invalid selection. Please enter a valid number.")
-                    logger.log_entry(username or "system", "Backup Menu", f"Invalid delete selection: {sel}", "No")
-                    continue
-                sel_idx = int(sel) - 1
-                try:
-                    delete_backup(backups[sel_idx], username)
-                    print(f"Backup deleted: {backups[sel_idx]}")
-                    logger.log_entry(username or "system", "Backup Menu", f"Deleted backup: {backups[sel_idx]}", "No")
-                except Exception as e:
-                    print(f"Delete failed: {e}")
-                    logger.log_entry(username or "system", "Backup Menu", f"Delete failed: {e}", "Yes")
-                break
-        elif role == "superadmin" and choice == "5":
-            backups = list_backups()
-            if not backups:
-                print("No backups available.")
-                logger.log_entry(username or "system", "Backup Menu", "No backups available for restore code generation.", "No")
-                continue
-            while True:
-                print("Available backups:")
-                for idx, b in enumerate(backups, 1):
-                    print(f"{idx}. {b}")
-                sel = input("Select backup number: ").strip()
-                if not sel.isdigit() or int(sel) < 1 or int(sel) > len(backups):
-                    print("Invalid selection. Please enter a valid number.")
-                    logger.log_entry(username or "system", "Backup Menu", f"Invalid backup selection for code: {sel}", "No")
-                    continue
-                sel_idx = int(sel) - 1
-                admins = get_system_admins()
-                if not admins:
-                    print("No active System Admins found.")
-                    logger.log_entry(username or "system", "Backup Menu", "No active System Admins found for code generation.", "No")
-                    break
-                while True:
-                    print("System Admins:")
-                    for idx, admin in enumerate(admins, 1):
-                        print(f"{idx}. {admin}")
-                    admin_sel = input("Select System Admin number: ").strip()
-                    if not admin_sel.isdigit() or int(admin_sel) < 1 or int(admin_sel) > len(admins):
-                        print("Invalid selection. Please enter a valid number.")
-                        logger.log_entry(username or "system", "Backup Menu", f"Invalid System Admin selection for code: {admin_sel}", "No")
-                        continue
-                    admin_idx = int(admin_sel) - 1
-                    try:
-                        code = add_restore_code(backups[sel_idx], admins[admin_idx])
-                        if code:
-                            print(f"Restore code for {admins[admin_idx]} and backup {backups[sel_idx]}: {code}")
-                            logger.log_entry(username or "system", "Backup Menu", f"Generated restore code for {admins[admin_idx]} and backup {backups[sel_idx]}", "No")
-                        break
-                    except Exception as e:
-                        print(f"Failed to generate restore code: {e}")
-                        logger.log_entry(username or "system", "Backup Menu", f"Failed to generate restore code: {e}", "Yes")
-                break
-        elif role == "superadmin" and choice == "6":
-            backups = list_backups()
-            if not backups:
-                print("No backups available.")
-                logger.log_entry(username or "system", "Backup Menu", "No backups available for revoke.", "No")
-                continue
-            while True:
-                print("Available backups:")
-                for idx, b in enumerate(backups, 1):
-                    print(f"{idx}. {b}")
-                sel = input("Select backup number: ").strip()
-                if not sel.isdigit() or int(sel) < 1 or int(sel) > len(backups):
-                    print("Invalid selection. Please enter a valid number.")
-                    logger.log_entry(username or "system", "Backup Menu", f"Invalid backup selection for revoke: {sel}", "No")
-                    continue
-                sel_idx = int(sel) - 1
-                admins = get_system_admins()
-                if not admins:
-                    print("No active System Admins found.")
-                    logger.log_entry(username or "system", "Backup Menu", "No active System Admins found for revoke.", "No")
-                    break
-                while True:
-                    print("System Admins:")
-                    for idx, admin in enumerate(admins, 1):
-                        print(f"{idx}. {admin}")
-                    admin_sel = input("Select System Admin number: ").strip()
-                    if not admin_sel.isdigit() or int(admin_sel) < 1 or int(admin_sel) > len(admins):
-                        print("Invalid selection. Please enter a valid number.")
-                        logger.log_entry(username or "system", "Backup Menu", f"Invalid System Admin selection for revoke: {admin_sel}", "No")
-                        continue
-                    admin_idx = int(admin_sel) - 1
-                    try:
-                        if revoke_restore_code(backups[sel_idx], admins[admin_idx]):
-                            print(f"Restore code for {admins[admin_idx]} and backup {backups[sel_idx]} revoked.")
-                            logger.log_entry(username or "system", "Backup Menu", f"Revoked restore code for {admins[admin_idx]} and backup {backups[sel_idx]}", "No")
-                        else:
-                            print("No active restore code found to revoke.")
-                            logger.log_entry(username or "system", "Backup Menu", f"No active restore code found to revoke for {admins[admin_idx]} and backup {backups[sel_idx]}", "No")
-                        break
-                    except Exception as e:
-                        print(f"Failed to revoke restore code: {e}")
-                        logger.log_entry(username or "system", "Backup Menu", f"Failed to revoke restore code: {e}", "Yes")
-                break
-        elif (role == "superadmin" and choice == "7") or (role != "superadmin" and choice == "5"):
-            logger.log_entry(username or "system", "Backup Menu", "Exited backup menu.", "No")
-            return
-        else:
-            print("Invalid choice.")
-            logger.log_entry(username or "system", "Backup Menu", f"Invalid menu choice (else): {choice}", "No")
 
 # === MAIN MENU BEFORE LOGIN ===
 def pre_login_menu():
